Final Post Processing Work
Fix the post processor to properly pull in values from the configuration file when populating the APPX Manifest.
Pull in the generated publisher certificate and add it as the correct signing cert for the package.
Add the Invoke-MsBuild module which allows us to build from the setup script without depending on having access to msbuild.
Removing the checked in XboxServices.config file.