XboxLivePostProcessing - Remove dependency on Visual Studio Tools for Unity
This post-processing step used a functionality of Visual Studio Tools for Unity for which there is an equivalent in Unity (at least in 2019.4.3f1).
By removing this dependency, it removes the obligation of having Visual Studio Tools for Unity installed on the machine building the project using Xbox Live plugin.
I moved the replaced functionality from XboxLivePostProcessing.cs to XboxLivePostProcessingUnityProject.cs.